refactor : add index export of `user` API methods
