+++ src/shared/api/index.ts
@@ -1,3 +1,4 @@
 export * from './axios';
 export * from './requests';
 export * from './list';
+export * from './user';
