+++ src/widgets/room/api/join.ts
@@ -1,7 +1,6 @@
 import { useMutation } from '@tanstack/react-query';
 
-import { REQUEST } from '@/shared/api';
-import { userPost } from '@/shared/api/user';
+import { REQUEST, userPost } from '@/shared/api';
 import { useModal } from '@/shared/hook';
 import { MODAL } from '@/shared/constants';
